Remove `handler.go`.
This was a leftover file from previous experimentation. The idea was to
not have a direct internal function `subscribe/1`, but instead define
functions of the form `on_*_event`. There would be machinery in place,
outside of `gojq`, to listen specifically to events, and call the
functions as the relevant event occurs. The upside is that this makes it
easy to have an alternative version of `i3jq` written entirely in bash
or similar, since no internal functions are needed: at the beginning of
the pipeline we put in the layout tree and at the end we get a string
representing commands.

However, I decided to scrap this because having an internal function for
`subscribe` makes for way simpler code and more immediately
understandable modules.

This module may still come in handy at some point, so here it is
committed to history.