Add machinery to auto-generate code for builtins.
Here, I am heavily relying on existing code in itchyny's gojq. I see
that there is some `go generate` magic going on there, but seeing as a
`Makefile` is already used anyway, I'm not sure of the added value (and
it seems to need more manual work) so I'm skipping it for now.
Undoubtedly there are better ways to do things, so comments are welcome.