`focus` is now iterative, with changed parameter.
When `focus` took a parameter, it used to be a number $n, taken to mean
"instead of the first item in the focus list, take the $nth". And you'd
only take one step down the tree.

But that's a bit of a niche use. It's probably more intuitive to have
`focus` mean: find the most focused node, ie, continue focusing until
you arrive at the actual *focused* container. ... and if you
parameterize it, then it becomes: keep focusing until the given
condition holds. That would mean that getting the focused workspace is
just `focus(.type == "workspace")`, the focused output is just
`focus(.type == "output")`... It means that we can get rid of some
builtins, while improving the readability of jq code that uses the
module. The original `focus` can then be moved to `focus_step`.