Pre-commit hook should only abort commits.
I don't like pre-commit hooks changing a commit under the hood. For one,
I make loads of mistakes and I'd like to spot them. (For example, the
previous hook would generate builtins based on the current working tree,
not the currently staged builtins, and so you might get builtins that
actually *do not* correspond with the committed changes to the
underlying jq files, which is exactly what it's trying to prevent.)

This commit starts with that: the pre-commit hook will no longer just
generate something and slap it onto the commit. It only detects when
it's unhappy with whatever is currently staged, and leaves the rest for
you to figure out.