Don't search `$ORIGIN/../share/i3jq/` for modules.
As of 4627aa48154be4b62ddc3baa55f55b6651545916, we assume that `i3jq` is
a single binary, so there's no point in looking for extra installed
files. It's either a builtin or something from `contrib/` that the user
should copy to their config if they want to. No in-between. This makes
it clear what can be expected from the `contrib/` modules.

I have removed the install script for `contrib/` scripts, accordingly.