Only load base module if writing query on the CLI.
When loading filters from a file, users should be explicit about what
they are importing, so loading the base module is unnecessary.

This also means that `ipc::get_tree`, which used to be ran when neither
a module nor a query was given, is now simply defined in the base
module. This is simpler (and also works with `i3jq -m base`).

Also cheekily snuck in module descriptions for each builtin.