Make `show.jq` a builtin, and make it the default.
In scripts, it is more acceptable to fiddle with switches, but when
running in the command line, you just want to get back something
readable when you type the minimal command `i3jq`. The visual tiling
tree is much easier to read (though it will need some cosmetic tweaking
still, as well as the addition of marks and such).

Bonus: now we always load a base module, and the `base.jq` module (which
was somewhat boring) can just be removed.