Rename to `i3q`
The brevity of the name is attractive. More importantly, I think
`swayjq` is not a nice name, whereas `swayq` is. I originally thought
the name should refer to the 'original' WM, even if I personally use it
with Sway, but on second thought, maybe it's better to treat Wayland as
the default. The lack of `jq` in the name was another reason I did not
make this change earlier. But who cares --- it's still obvious.