Start rework of master-stack layout.
The master-stack layout script needs an update. My initial thoughts for
this rework are as follows:

- It should make no assumptions on the current layout tree in order to
  function as expected.
- The master window must not be in a static position. It can be placed
  at the top, bottom, left or right of the window.
- The orientation should be changeable on-the-fly.
- Mouse actions should be able to change the orientation.
- On initialization, either all workspaces should be put in this layout,
  or it should be possible to dynamically turn the layouting on for
  specific workspaces.
- Perhaps there should be an arbitrary number of master windows/an
  arbitrary number of stack containers, to accommodate wide monitors.

Not all of these ideas may be immediately implemented in this branch. I
might postpone or rework or entirely abandon some ideas.

This commit rearranges the code somewhat to be more in line with how I
want to approach this.