Use numerical position for overflow nodes.
Even though it's slightly more complicated to understand in some
respects, it also eliminates some conditionals. It is more generic and
has fewer lines of code. The actual reason I chose to do it this way,
however, is that we might, in the future, want to support having
multiple overflow nodes, and in that case this structure makes our life
a whole lot easier.