Rename binary (and project itself) to `swayq`.
In 059dc7cff40addfe985be34af4fb621c487f429c, I was already toying with
this (as evidenced by the confused babble in that commit message). I
like that I can use this project as a low-stakes way to indulge in
indecision --- get that out of my system. In any case, this should be
the last name change!