Simpler tiling description with multiple overflows.
The description of tiling layouts is now simpler and more general, but
it comes at the cost of more complicated internal logic.

The idea is to allow multiple overflow windows, which can accommodate
much more flexible layouts, such as a layout with two stacks (for a
wide-screen monitor). It also comes with a more intuitive way to define
layouts: you describe subschemas with a certain priority, and the module
figures out the rest.

This loses some functionality, like inheriting root schemas, which are
to be reinstated later.