Set `$ARGS` depending on query content.
If a module only sets function definitions, it makes sense to treat the
first non-flag argument as a filter to execute. But if the module
already provides a filter, then all non-flag arguments are our `$ARGS`.

Perhaps this is a bit less intuitive for the writer of filters, but it
makes for a cleaner interface for the user of the CLI (since you don't
necessarily need to provide a filter before parsing your `$ARGS` --- you
can do `swayq module --argument` instead of `swayq module main_filter
--argument`).