Add `modules` builtin for index and completions.
The `modules` builtin simply outputs all available modules. We can use
the already existing `modulemeta` to get information on the defined
functions and metadata on each module, and so we can use it for both
completions and the module overview.

This is more elegant and makes some Go code redundant.